Modifications
Some modifications are made. The training has not finished yet, but the FID is decreasing (though slowly). n_critic is chosen to be 5 in the current case.

The learning rate is tested to be at most 0.0001 to support stable training.

A new way to implement WANG-GP loss is added, which calls backward() multiple times therefore requires less memory. However, it is not the only bottle neck for memory usage, so the implementation actually used is the more concise version.

A bug in GP loss is fixed: the dim for computing the 2-norm should be (1,2,3).

Add the resuming training option. Training can be resumed for the latest epoch.

Raise the weight of statistical constraint (spanwise mean velocity) from 10 to 100, and better performance is observed.

Add the standard deviation layer to the Discriminator.

Add augmentation of dataset: flip in spanwise direction.