stablize training with normalization
Add the Equalized learning rate (Karras et al. 2018, codes adapted from https://github.com/rosinality/progressive-gan-pytorch) feature to all Conv2d and Linear layers in D and G. This operation effectively rescales learning rate for each layer during every update, and therefore greatly stablizes the training. Learning can be chosen 0.001 instead of 0.0001.

Rearrange the D and G networks into a models.py script so that they can shear common classes/functions such as EqualLR and conv_with_padding.

Add a statistical constraint to restrict the variations across Z-direction and batch to be close to that of real flow fields. The constraint is not strict (Yang, Wu & Xiao 2021) because the Reynolds numbers are different for samples in a single batch.

The current configuration and parameters leads to a relative FID 3.65 (still slowly decreasing). The current version is committed as a checkpoint before further tuning of the networks and parameters.