Slight modifications
Change the inputs for constructing Generator and Discriminator. Instead of an 'options' packing all parameters, the needed parameters are now explicitly required as __init__ arguments.

Combine the inputs of the two terms in WGAN loss into one input tensor so that the loss can be computed by going through D net once. This is expected to be faster because a larger tensor make more use of GPU, but sometimes it may cause a GPU memory error.

Add a CUDA switch to conveniently turn on/off CUDA. (the device of inception net in FID modified accordingly)