Apply Residual blocks successfully
The order of layers in repeated blocks are adjusted to fit the structure of CcGAN (https://github.com/UBCDingXin/improved_CcGAN). The Generator is equivalent to the original one after adjustment, but the Discriminator differ from the original one in that the LReLU and downsampling are swapped.

Residual net is applied by connecting the two ends of each convolutional block, following the new version of ResNet proposed by He et al. (2016), that is, 'norm-act-conv-norm-act-conv'.

The new structure has been tested and good results (convergence on FID 4) has been obtained.