Major reconstruction
Split dcgan.py into a main.py and a train.py, respectively control the executing flow and the training process.

Modify the models to version2 and version3. New models are tested to converge to satisfying FID (generally under 5), but still need to explore more reasonable structures and training parameters because the best performance so far is FID = 1~2.

Default arguments are added to the model inputs to handle both unsupervised and supervised training.

The losses are modified to take and give suitable arguments for the models. And every term in the loss is output so that the training process can be better monitored.