Multi-GPU training
Use Torch.distributed to realize multi-GPU training.

A bug in models fixed: in Generator, the getone() method should generate a latent vector with distribution randn() instead of rand().

In the MyDataset, in order to build a sufficiently large dataset, the dataset is generated from the raw data every second streamwise points, instead of every forth streamwise points.

Also in MyDataset, some functions in __getitem__ are moved to the get_sample() method.