Model and training improved. Training succeeded!
Fix the bug in models. Change the in/out feature numbers so that the Discriminator is truly the same to that in Ding et al.

Wrap the weighted random sampler into the dataset, set number of samples as batch_size in each draw. The sampling is critical to the outcome of training!

Add some ways of learning rate decay, and use config parameters to control them.